+++ assets/translate/groupingstrategy.json
@@ -1,6 +1,6 @@
 {
     "BY_MARKET": "Group By Market",
-    "BY_SYSTEM": "Group By System",
+    "BY_SYSTEM": "Group By Star System",
     "PER_LOCATION": "Per Location",
     "UNIFIED": "Unified"
 }
