+++ assets/translate/query.json
@@ -19,9 +19,11 @@
     "MANAGE_DMOD_COUNT": "D-mod count",
     "MANAGE_DMOD_SET": "Allowed and disallowed d-mods",
     "MANAGE_FILTERING": "Market and submarket filtering",
+    "MANAGE_FILTERING_OTHER": "Other filters",
+    "MANAGE_FILTERING_SUBMARKET": "Include submarkets",
     "MANAGE_GROUPING_SORTING": "Result grouping and sorting",
-    "MANAGE_RESULT_GROUPING": "Result grouping",
-    "MANAGE_RESULT_SORTING": "Result sorting",
+    "MANAGE_RESULT_GROUPING": "Grouping method",
+    "MANAGE_RESULT_SORTING": "Sorting criteria",
     "MANUFACTURERS": "Design type",
     "MATCHING_ITEMS": "Matching items",
     "MATCHING_SHIPS": "Matching ships",
