+++ src/stelnet/board/query/QueryL10n.java
@@ -21,6 +21,8 @@ public enum QueryL10n {
     MANAGE_DMOD_COUNT,
     MANAGE_DMOD_SET,
     MANAGE_FILTERING,
+    MANAGE_FILTERING_OTHER,
+    MANAGE_FILTERING_SUBMARKET,
     MANAGE_GROUPING_SORTING,
     MANAGE_RESULT_GROUPING,
     MANAGE_RESULT_SORTING,
